register product page added

+++ client/src/appConfig/app.config.js
@@ -5,22 +5,22 @@
 'use strict';
 
 // Default environment variables
-  var _env = {};
-  
-  // Importing variable
-  if(window){
+var _env = {};
+
+// Importing variable
+if (window) {
     Object.assign(_env, window.__env);
-  }
+}
 
 
 angular
-    .module('appConfig', ['LocalStorageModule', 'ngResource', 'ui.bootstrap', 'ngTable', 'ngAnimate', 'ngSanitize', 'ngFileUpload','angularjs-dropdown-multiselect'])
+    .module('appConfig', ['LocalStorageModule', 'ngResource', 'ui.bootstrap', 'ngTable', 'ngAnimate', 'ngSanitize', 'ngFileUpload', 'angularjs-dropdown-multiselect'])
     .config(['$httpProvider', '$logProvider', function ($httpProvider, $logProvider) {
 
         // CORS (Cross-Origin Resource Sharing) headers to support Cross-site HTTP requests
         $httpProvider.defaults.useXDomain = true;
         delete $httpProvider.defaults.headers.common['X-Requested-With'];
-        
+
         //Configure http interceptor and application loader
         $httpProvider.interceptors.push('httpInterceptorService');
 
@@ -41,15 +41,15 @@ angular
                     return response || $q.when(response);
                 },
                 responseError: function (response) {
-                    try{
+                    try {
                         $rootScope.hasError = true;
-                        if(appConstants.ACCESS_DENIED_CODE.indexOf(response.status) >= 0){
+                        if (appConstants.ACCESS_DENIED_CODE.indexOf(response.status) >= 0) {
                             $rootScope.ERROR_MSG = appConstants.UNAUTHORIZED_ERROR;
-                        }else{
+                        } else {
                             $rootScope.ERROR_MSG = appConstants.SERVICE_ERROR;
                         }
                         $rootScope.$broadcast("loaderHide");
-                    }catch(e){
+                    } catch (e) {
                         $log.error(appConstants.FUNCTIONAL_ERR, e);
                     }
                     return $q.reject(response);
@@ -72,14 +72,31 @@ angular
     })
 
     .run(['$rootScope', '$window', 'localStorageService', '$log', function ($rootScope, $window, localStorageService, $log) {
-        
+
         $log.debug('appConfig bootstrapped!');
 
         $rootScope.isLoggedIn = false;
         $rootScope.activeMenu = '';
         $rootScope.hasError = false;
         $rootScope.ERROR_MSG = "";
-        $window.onunload = function () {
-            localStorageService.remove('User');
-        };
+        $window.onunload = callbackFunction(localStorageService);
+
     }]);
+
+function callbackFunction(localStorageService) {
+
+    if (window.event) {           
+        if (window.event.clientX < 40 && window.event.clientY < 0) {
+            console.log("not refreshed");   
+            localStorageService.remove('User');
+        } else {
+            
+        }
+    } else {
+        if (event.currentTarget.performance.navigation.type == 2) {
+            localStorageService.remove('User');
+        }
+        if (event.currentTarget.performance.navigation.type == 1) {
+        }
+    }
+};
